Add mobile build directory
