+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "extensions",
     "native_extensions/emoji_selector",
 ]
+exclude = ["mobile"]
 
 [profile.rapid]
 inherits = "dev"
@@ -26,17 +27,17 @@ version = "0.1.5"
 rust-version = "1.70"
 
 [workspace.dependencies]
-wry = { version = "0.27.2" }
-dioxus = { git = "https://github.com/DioxusLabs/dioxus", rev = "365869806401655851f219e325d0d1a9e0dad7e4" }
-dioxus-hooks = { git = "https://github.com/DioxusLabs/dioxus", rev = "365869806401655851f219e325d0d1a9e0dad7e4" }
-dioxus-html = { git = "https://github.com/DioxusLabs/dioxus", rev = "365869806401655851f219e325d0d1a9e0dad7e4" }
-dioxus-router = { git = "https://github.com/DioxusLabs/dioxus", rev = "365869806401655851f219e325d0d1a9e0dad7e4" }
-dioxus-desktop = { git = "https://github.com/DioxusLabs/dioxus", rev = "365869806401655851f219e325d0d1a9e0dad7e4", features = [
+wry = { version = "0.28" }
+dioxus = { git = "https://github.com/DioxusLabs/dioxus" }
+dioxus-hooks = { git = "https://github.com/DioxusLabs/dioxus" }
+dioxus-html = { git = "https://github.com/DioxusLabs/dioxus" }
+dioxus-router = { git = "https://github.com/DioxusLabs/dioxus" }
+dioxus-desktop = { git = "https://github.com/DioxusLabs/dioxus", features = [
     "transparent",
 ] }
 raw-window-handle = "0.5"
-dioxus-core = { git = "https://github.com/DioxusLabs/dioxus", rev = "365869806401655851f219e325d0d1a9e0dad7e4" }
-fermi = { git = "https://github.com/DioxusLabs/dioxus", rev = "365869806401655851f219e325d0d1a9e0dad7e4" }
+dioxus-core = { git = "https://github.com/DioxusLabs/dioxus" }
+fermi = { git = "https://github.com/DioxusLabs/dioxus" }
 tokio-util = { version = "0.7", features = ["full"] }
 arboard = "3.2"
 humansize = "2.1.3"
@@ -81,7 +82,9 @@ zip = "0.6.4"
 
 image = "0.24"
 linkify = "0.9.0"
-reqwest = "0.11.16"
+reqwest = { version = "0.11.16", features = [
+    "rustls-tls",
+], default-features = false }
 select = "0.6.0"
 
 common = { path = "common" }
