+++ common/Cargo.toml
@@ -15,7 +15,7 @@ serde.workspace = true
 serde_json.workspace = true
 once_cell.workspace = true
 clap.workspace = true
-rodio = "0.16"
+rodio = "0.17"
 uuid.workspace = true
 notify-rust.workspace = true
 dirs.workspace = true
