+++ mobile/.gitignore
@@ -1,0 +1,10 @@
+# Rust
+target/
+**/*.rs.bk
+
+# tauri-mobile
+.cargo/
+/gen
+
+# macOS
+.DS_Store
