+++ mobile/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "mobile"
+version = "0.1.0"
+authors = ["Jonathan Kelley <jkelleyrtp@gmail.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["staticlib", "cdylib", "rlib"]
+
+[[bin]]
+name = "mobile-desktop"
+path = "gen/bin/desktop.rs"
+
+[package.metadata.cargo-android]
+app-activity-name = "com.example.mobile.MainActivity"
+app-dependencies = [
+    "androidx.webkit:webkit:1.6.1",
+    "androidx.appcompat:appcompat:1.6.1",
+    "com.google.android.material:material:1.8.0",
+]
+project-dependencies = ["org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.21"]
+app-plugins = ["org.jetbrains.kotlin.android"]
+app-permissions = ["android.permission.INTERNET"]
+app-theme-parent = "Theme.MaterialComponents.DayNight.DarkActionBar"
+vulkan-validation = false
+
+[package.metadata.cargo-android.env-vars]
+WRY_ANDROID_PACKAGE = "com.example.mobile"
+WRY_ANDROID_LIBRARY = "mobile"
+WRY_ANDROID_KOTLIN_FILES_OUT_DIR = "<android-project-dir>/app/src/main/kotlin/com/example/mobile"
+
+[package.metadata.cargo-apple.ios]
+frameworks = ["WebKit"]
+
+[dependencies]
+anyhow = "1.0.56"
+log = "0.4.11"
+wry = "0.28.0"
+uplink = { path = "../ui" }
+
+[target.'cfg(target_os = "android")'.dependencies]
+android_logger = "0.9.0"
+jni = "0.19.0"
+paste = "1.0"
+
+[target.'cfg(not(target_os = "android"))'.dependencies]
+env_logger = "0.9.0"
+
+[target.'cfg(target_os = "ios")'.dependencies]
+core-foundation = "0.9.3"
