+++ ui/Cargo.toml
@@ -18,7 +18,7 @@ kit.workspace = true
 common.workspace = true
 extensions.workspace = true
 tokio-util.workspace = true
-arboard.workspace = true
+# arboard.workspace = true
 humansize.workspace = true
 window-vibrancy.workspace = true
 uuid.workspace = true
